Fix lint warnings
