+++ .github/workflows/azure-static-web-apps-CI.yml
@@ -22,9 +22,9 @@ jobs:
         run: |
           npm install
 
-#      - name: Linting
-#        run: |
-#          npm run lint
+      - name: Linting
+        run: |
+          npm run lint
 
       - name: Run unit tests
         run: |
